Use the sniff properties to exclude some errors rather than excluding by message.
This will make the running of the sniffs more efficient (faster) as it will skip the rest of the logic in the sniff for excluded groups.

For the `WordPress.VIP.RestrictedVariables` and `WordPress.VIP.PostsPerPage`, the current excluded groups effectively disable the sniffs, however, excluding the sniffs completely would prevent any potentially new groups being added in the future from being recognized.