:rocket: Fix travis unit test failure
The Travis PHP images for PHP 7+ have started to use PHPUnit 6, while the PHPCS testsuite is not yet compatible with that.

This commit is intended (fingers crossed) as a temporary solution to be reverted once the testsuite for PHPCS is cross-version compatible with PHPUnit.

I've chosen to only download PHPUnit when a different version is needed than the one installed by default in the Travis images to keep the build time down.