Lint the XML files as part of the build testing
Checks the syntax of XML files + the layout.
In line with WP, tabs are specified to be used for indentation, not spaces.

I've chosen to run this during one of the PHP 5.3 builds so as not to have to split of another PHP version from the main matrix. The PHP version should have no effect on the results anyhow, so this should be fine.