Composer: Up the required PHPCS and WPCS versions
The latest PHPCS release is `3.1.1` which came out mid October. The minimum PHP requirement for PHPCS 3.x is PHP 5..4.
The latest WPCS release is `0.14.0` which has come out November 1st.

This commit also updates the recommended version of the DealerDirect PHPCS Composer to be in line with the recommendation included with WPCS 0.14.0.

Also: Be less specific about the minimum requirement PHPCS version.

All WPCS sniffs, as well as the sniff in this repo are tested again multiple PHPCS versions between the minimum requirement and the maximum supported version, so using a higher PHPCS version should be fine.

The changes between versions of WPCS are far more far reaching, the effect of newer PHPCS versions is largely bugfixes, so being less specific will positively benefit YoastCS.