Update YoastCS native custom ruleset & build for PHPCS 3.3.0
As of PHPCS 3.3.0, a configuration directive set in a custom ruleset, can be overruled from the command-line.

This allows us to get rid of the excludes in the YoastCS own ruleset which were needed as the minimum PHP requirement for YoastCS 0.5+ is PHP 5.4, while WP has a minimum requirement of PHP 5.2.4.

This also adds a `check-cs` script to the composer file to make it easier for people to run phpcs locally with the right command-line arguments.