Travis: fix build failure
The current PHP 7.2/nightly images on Travis come by default with PHPUnit 7.x.

Unfortunately, the lowest PHPCS version supported by YoastCS (`3.2.0`), did not yet have full support for PHPUnit 7.x. This has since been fixed, which is why the builds against PHPCS `master` are passing fine.

To still be able to test the sniffs in YoastCS fully, we need to use PHPUnit 6.x for the PHP 7.2/nightly builds in combination with PHPCS 3.2.0, so that's what this PR sorts out.