Composer: add note about why the DealerDirect plugin is suggested
Just suggesting a package does not explain to a user why it would be interesting for them to install it.

The `suggest` entry in the `composer.json` file has a value which is not actually intended for version constraints, but to add a hint to the user about why they should install the package.

Ref: https://getcomposer.org/doc/04-schema.md#suggest