Use PHP_CodeSniffer v 3.3.0
This updates YoastCS to use PHPCS 3.3.0 which was released this week.
See: https://github.com/squizlabs/PHP_CodeSniffer/releases/tag/3.3.0

The most significant change in the context of YoastCS is that `array` type properties set in the ruleset now use a new format. The old format is still supported for now, but will be removed in PHPCS 4.0.