Ruleset: add two new sniffs
The `LowerCaseType` sniff verifies that all type casts, parameter and return type declarations are in lowercase (with the exception of classname based type declarations).
The `ShortFormTypeKeywords` sniff ensures that the same type indicators are used in short form, i.e. `int`, not `integer`.

I don't expect any new errors to be thrown up because of these sniffs as what they check is common practice, but, just in case, it will now be enforced.

These sniffs are new to PHPCS 3.3.0 and will be included in WPCS once the WPCS minimum PHPCS version will be upped to PHPCS 3.3.0. This is expected to happen in WPCS 2.0. So, once YoastCS moves onto WPCS 2.0, these additions can be removed.