Remove the composer.lock file
Committing the `composer.lock` file for libraries is not necessary as it will be ignored anyway by the consumers of the package.

Also, as this library is supposed to support various versions of PHPCS and in the future possibly various version of WPCS, it is actually a bad idea to lock the versions as this makes testing against different PHPCS/WPCS combinations more involved.