Start using PHPCompatibilityWP
PHPCompatibility has been moved to a dedicated GitHub organisation account and has released version 8.2.0.

A `PHPCompatibilityWP` ruleset is now also available in a separate repo which already excludes all back-fills/poly-fills which WP provides.

This ruleset should from now on be used to test the plugins.

For testing this repository, the "normal" PHPCompatibility ruleset should still be used as the code from this repo is not run within the context of WP, but in the context of PHPCS.

For more details, see the [release notes of the PHPCompatibility 8.2.0 release](https://github.com/PHPCompatibility/PHPCompatibility/releases/tag/8.2.0).