:sparkles: New sniff to check that test double/mock classes are in the correct directory
This sniff looks for classes containing `Mock` or `Double` (case-insensitive) in the class name and will:
1. Check that a `tests/doubles` directory exist and throw an `error` if it doesn't.
2. Check that the file containing the double/mock class is within the `tests/doubles` directory and throw an `error` if it isn't.
3. Check that the double/mock class is in a file of its own and throw an `error` if it isn't.

The `doubles_path` by default is `/tests/doubles` relative to the `basepath` for the project.
This can be customized from within a custom ruleset by passing a different value for `doubles_path`.
```xml
<rule ref="Yoast.Files.TestDoubles">
	<properties>
		<property name="doubles_path" value="/tests/assets"/>
	</properties>
</rule>
```

For this sniff to function, the `basepath` for the project has to be set. If it's not a `warning` will be thrown.
The basepath will normally be set from within the custom ruleset like so:
```xml
<arg name="basepath" value="./"/>
```

Includes unit tests covering all errors and warnings.

As YoastCS itself does a) not contain mocks/doubles and b) has to follow the predetermined directory structure as set by PHPCS, for the YoastCS native CS check, this sniff is disregarded (excluded).