Composer: use PHPCS 3.3.1
Aside from it being a new release with lots of bug fixes, there are also a few specific changes in PHPCS 3.3.1 which benefit YoastCS and necessitate updating the minimum version:

> * Generic.Commenting.DocComment no longer generates the SpacingBeforeTags error if tags are the first content in the docblock
>     - The sniff will still generate a MissingShort error if there is no short comment
>     -This allows the MissingShort error to be suppressed in a ruleset to make short descriptions optional

As a lot of plugins still have issues with missing short descriptions, the `SpacingBeforeTags` error was also being thrown and had to be forcibly excluded. These exclusions can now be removed.

> * Squiz.WhiteSpace.FunctionSpacing no longer checks spacing of functions at the top of an embedded PHP block
>     - Fixes a conflict between this sniff and the Squiz.PHP.EmbeddedPHP sniff

This bug was discovered via one of the Yoast plugins, so having this fix prevents needing a work-around.

Ref: https://github.com/squizlabs/PHP_CodeSniffer/releases/tag/3.3.1