Add a changelog for the YoastCS package
A well-maintained changelog makes it easier for the "product-owners" of each Yoast plugin to upgrade to a higher version of YoastCS, as it is documented what has changed in each version.

As no changelog existed until now, this PR adds a changelog for all YoastCS releases since the very first one.

Fixes 77

To do:

- [ ] Add the changelogs for each release to the GH release tags.