:sparkles: New `CodeCoverageIgnoreDeprecated` sniff
This sniff will check that functions which have a `@deprecated` tag in the function docblock, also have a  `@codeCoverageIgnore` tag in the same docblock.

If a `codeCoverageIgnore` text is found without the `@` sign, it will be recognized by the sniff and can be auto-fixed.

Individual plugins may have certain files `exclude`d for codeCoverage in the PHPUnit configuration file.
For those plugins a corresponding `exclude`s needs to be added to the PHPCS configuration file to ignore those files completely for the purposes of this sniff.

Example:
```xml
<rule ref="Yoast.Commenting.CodeCoverageIgnoreDeprecated">
    <exclude-pattern>*/deprecated/*\.php$</exclude-pattern>
</rule>
```