Update PHPCS and PHPCompatibility dependencies
[PHPCompatibility 9.0.0](https://github.com/PHPCompatibility/PHPCompatibility/releases/tag/9.0.0) has just been released and renames all sniffs.
[PHPCompatibilityWP 2.0.0](https://github.com/PHPCompatibility/PHPCompatibilityWP/releases/tag/2.0.0) has been released alongside it.

Asides from that [PHP_CodeSniffer 3.3.2](https://github.com/squizlabs/PHP_CodeSniffer/releases/tag/3.3.2) has also been released with improved report layouts and a number of bugfixes.

The upgrade for this repo is simple as YoastCS currently doesn't use PHPCS inline annotations which reference PHPCompatibility sniffs, nor uses custom excludes.

As the PHPCompatibilityWP ruleset as of version 2.0 has an additional dependency, it is however opportune to let the DealerDirect plugin sort out the installed paths, at least for devs.

For the users of YoastCS, the DealerDirect plugin is advised anyway and for all repos which are already using a recent version of YoastCS, this has been implemented.