:rocket: Build/Travis: Validate the PHPCS ruleset against XSD schema
As of PHPCS 3.2.0, PHPCS includes an XSD schema for rulesets which defines what can be used in the XML ruleset file.

In PHPCS 3.3.0 a new array format for properties was introduced and as of PHPCS 3.3.2, the new array format is now also covered by this schema.

PR 99 updated the PHPCS requirement to PHPCS 3.3.2, so now we can actually use the XSD and validate against it.

This commit:
* Adds the schema tags to the `ruleset.xml` file as well as the native YoastCS `phpcs.xml.dist` file.
* Adds validation against the schema for the `ruleset.xml` file to the Travis build.

Fixes 91