Introduces PHPUnit bootstrap logic
Because we use WPCS 1.x, which still supports PHPCS 2.x, we need to make sure the WPCS alias file which adds cross-version support for PHPCS 2.x and 3.x is loaded, when running the unit tests.

As this introduces a YoastCS PHPUnit bootstrap file, this file should also take case of loading the PHPCS 3.1+ PHPUnit bootstrap which sorts out PHPUnit cross-version compatibility for PHPCS.

Once WPCS 2.0 has been released and bumped, this workaround can be removed.