Composer: Update DealerDirect Composer plugin and move to require
The [DealerDirect Composer plugin](https://github.com/Dealerdirect/phpcodesniffer-composer-installer/releases) has released version 0.5.0.

I've also done some testing to see if when the plugin is in the YoastCS `require` section, it will work, even if the individual projects which have YoastCS as a dependency do *not* explicitly require the DealerDirect plugin and have found this to work smoothly in most cases.

With that in mind, I've moved the dependency from the `require-dev` section to the `require` section and have removed  the `suggests` section.

This means that - in most cases - projects which `require(-dev)` YoastCS, will no longer need to have the plugin in their own `composer.json`.