TestDoubles: allow for multiple paths
As discussed with moorscode, in rare circumstances, a plugin should be able to allow more than one directory to be used to place test doubles/mocks in.

In this commit, the `doubles_path` sniff property is changed from a string to an array property.

In the previous version of this sniff, when the `doubles_path` would not be valid, both a `NoDoublesDirectory` as well as a `WrongDirectory` error would be thrown.

In this version, the `NoDoublesDirectory` error will only be thrown if none of the relative paths supplied exist.
If at least one path is valid and that path is not used, the sniff will only throw a `WrongDirectory` error.

Notes:
* Includes BC-break protection for if the value would still be supplied as a string to prevent BC-breaks.
    This snippet can be removed in YoastCS 2.0.0.
* Now uses PHPCS 3.2+ style annotations for the property changing within the test case files.
    This is preferred since setting array properties using `//phpcs:set` was greatly improved in PHPCS 3.3.0.
    See: squizlabs/PHP_CodeSniffer 1999

Includes unit tests.