TestDoubles: improve path comparison
Make sure target_path always has an end slash to prevent a path like `/tests/doublesniff` to be recognized as correct when `/test/doubles` is in the allowed list.

This also changes the comparison from just using the `dirname` of the file being examined to using to complete file name, to prevent those kind of mismatches.

Unlikely to be a problem in real life, but the code should be correct.