FileComment: Fix the tests
As the unit tests are run without the context of the `ruleset.xml` file, the exclusions made in that file will not be applied.

For that reason, the unit test files which test that the sniff correctly falls through to the Squiz parent sniff, need to comply with the full rules from the `Squiz.Commenting.FileComment` sniff, including the Squiz author/copyright tags.