PHPUnit: add Composer script to run the tests
.. and various other tweaks.

### Unit testing the sniffs

A `dev` requirement of `phpunit/phpunit` has been added to the `composer.json` file, as well as a script to run the unit tests.
While PHPUnit is not strictly speaking a dependency of YoastCS (but of PHPCS), this just makes life easier on YoastCS developers who use the Composer install for development.

If - as a YoastCS sniff developer - you've installed YoastCS using Composer, you can now run the unit tests, like so:
```bash
composer test
```

**Note**: the `travis.yml` script does not use this script as it uses the Travis image native PHPUnit version instead of a (potentially) Composer installed one. For most builds PHPUnit wouldn't be Composer installed anyway as - except for one - all builds use `--no-dev`.

### Use the PHP version used by Composer for Composer scripts

Using the `@php` prefix in the scripts ensures that the scripts are run against the same PHP version as with which Composer is called, instead of against the system default PHP version.

This is important if - as a YoastCS dev - you change the PHP version for Composer to be able to test things against different PHP versions.

Ref: composer/composer 7645

### Minor cleanup of the `PHPCS_IGNORE_TESTS` directive

WPCS 2.0.0 will remove the `WordPress-VIP` standard. As that standard, as well as the other sub-standards of WPCS don't have their own unit tests anyway, we may as well remove references to them from the `phpunit.xml.dist` `PHPCS_IGNORE_TESTS` environment variable.