PHPCS/PHPCompatibility: only scan PHP files
The YoastCS ruleset can be used to scan PHP, JS and CSS files, just like PHPCS and WPCS.

The PHPCompatibility ruleset, however, is only aimed at PHP files.

This small tweak makes sure that if a project contains JS and CSS files and the project based ruleset allows for those to be scanned, that the PHPCompatibility checks are only applied to PHP files.