Composer: sync with other config files
I've done a compare between the `composer.json` config files in (nearly) all plugin/library repos.

This commit adds some additional properties to the config file to improve consistency with the other repos.

Additionally, it adds a `fix-cs` script.

Note: The `fix-cs` script does not have `--runtime-set testVersion 5.4-` in the command as that is only relevant for PHPCompatibility and the PHPCompatibility standard does not contain any fixers, so setting the `testVersion` is not needed when running `phpcbf`.