Travis: work around PHPUnit 8.x on PHP >= 7.2 images
As of recently, the Travis images for PHP 7.2+ ship with PHPUnit 8.x.
The PHPCS native test framework is not compatible with PHPUnit 8.x and won't be able to be made compatible with it until the minimum PHP version would be raised to PHP 7.1.

So for the unit tests to be able to run on PHP 7.2+, we need to explicitly require PHPUnit 7.x for those builds.

As for nightly: there is no PHPUnit version which is currently compatible with PHP 8.

---

Technical explanation:

PHPUnit 8 adds `void` return type declarations to the PHPUnit methods used by PHPCS, making it neigh impossible to make the unit test suite cross-version compatible for the PHP versions supported by PHPCS as the `void` return type was only added in PHP 7.1 and the minimum PHP version supported by PHPCS is 5.4.

Officially, PHPUnit 7 is compatible with PHP 7.1, 7.2 and 7.3. However for the functionality used by the PHPCS test suite, it looks to be compatible with PHP 7.4 as well (for now).
Ref: https://phpunit.de/supported-versions.html

For this reason, for Travis images which come natively with PHPUnit 8 (PHP >= 7.2), PHPUnit 7 needs to be installed via Composer.
This is not necessary for older PHP versions as the Travis native PHPUnit version works fine for those.

As for nightly/PHP 8:

* The Travis native PHPUnit version for nightly is PHPUnit 8, which will not work.
* Running composer install for PHPUnit on nightly currently installs PHPUnit 4.1.6. Most likely because the PHPUnit 4.x composer.json file did not yet contain a PHP requirement, while any higher PHPUnit versions do. Anyways, that will most definitely not work.
* Running composer install for PHPUnit with an explicit requirement of PHPUnit 7.x will also not work as PHPUnit 7 will not install on PHP 8 based on the PHPUnit composer.json file.

All in all, the unit tests can not currently be run on PHP 8.