YoastCS/FileName: remove support for passing properties as strings
Both the public `exclude`, as well as the `prefixes` property for this sniff are by nature array properties.

Setting array properties within unit tests was not yet properly supported by PHPCS until version 3.3.0. For that reason, the unit tests for this sniff still used the old-style property setting annotation and the sniff itself contained code to work around the issue.

As this problem has been solved now in PHPCS, the work-arounds in the sniff can be removed and the new-style property setting can now be used in the unit tests.