YoastCS/FileName: always remove the longest matching prefix when prefixes overlap
The sniff allows for passing a set of `prefixes` which will be removed from an OO name to get to the expected file name.

As it was, when two `prefixes` passed would overlap - think `yoast` and `yoast_news` -, the order in which these were provided in the configuration file would determine the order in which the sniff would attempt to match these.

I.e. if the prefixes were provided in the above order and the class name was `Yoast_News_Something`, the sniff would suggest `news-something.php` as the filename, not `something.php`.

The change in this commit, changes that behaviour by ordering the `prefixes` array before using it which ensures that the _longest_ matching prefix will be used when two prefixes overlap.

Includes unit test.

Unfortunately, as the tests are more similar to integration tests rather than unit tests, the actual message is not tested, even though that's where the change lies.