PHPCS ruleset: add sniff to check spacing between property declarations
The `Yoast.WhiteSpace.FunctionSpacing` sniff which was added in 86 checks that each function within an OO structure has one blank line before the function declaration and no blank line(s) after the last function in a class.

This sniff was added as there was quite some inconsistency regarding this across classes and repos and this way we could enforce consistency.

Similar to this, we should also enforce consistency regarding the spacing out of property declarations in OO structures., which I've noticed becoming inconsistent in some repos as well.

The sniff I'm proposing to add is a PHPCS native sniff which checks that there is one blank line before the first property declaration in an OO structure, as well as one blank line between property declarations in OO structures.

The sniff include an auto-fixer.

Ref: https://github.com/squizlabs/PHP_CodeSniffer/wiki/Customisable-Sniff-Properties#squizwhitespacemembervarspacing