:sparkles: New `ObjectNameDepth` sniff
This new sniff checks that the names of declared objects - class, interfaces, traits - contains no more than three words / two underscores.

This rule is only applied to objects declared in a namespace. For non-namespaced objects the rule does not apply.

The sniff will throw a `warning` when the name consists of more words than `recommended_max_words` and an `error` when the name is longer than `max_words`.

Both `recommended_max_words`, as well as `max_words` are configurable and can be changed from a custom ruleset.

By default, `recommended_max_words` and `max_words` are set to the same value: `3`, i.e. three words.

Note: the sniff currently extends the WPCS `Sniff` class to allow for using one of the utility methods contained in WPCS. Once `PHPCSUtils` comes out and use of `PHPCSUtils` is implemented in YoastCS, this will no longer be necessary and the (improved) utility method in `PHPCSUtils` can be used instead.

Includes unit tests.
Includes documentation.

Fixes 142