FileName: rename public properties
Sniff properties are specific to individual sniffs, but can be set for all sniffs in one go (undocumented PHPCS feature).

The public `$prefixes` property is used by the WPCS `PrefixAllGlobals` sniff to pass on which text strings should be considered "prefixes" for namespaces, class names, hook names, variable names etc for the purposes of that sniff.

Both the new, upcoming `ValidNamespace` sniff (141) and the Yoast specific `ValidHookName` sniff (143) will need the `$prefixes` property as well and the content for the property for those sniffs should overlap with the `$prefixes` property for the `PrefixAllGlobals` sniff.

To prevent having to set the same value three times in each custom ruleset, it would be useful to be able to use the "_set the property for all sniffs in one go_" feature.

However, the YoastCS `FileName` sniff also uses a `$prefixes` property which is a distinct, different property which indicates which part at the start of a class name should be removed to get the file name. This property may partially overlap with the `$prefixes` property for the `PrefixAllGlobals` sniff, but not necessarily.
The overlap will become smaller/non-existent once all plugins will start using namespaces.

To prevent confusion between the two types of `$prefixes` properties, I'm changing the property name in the `FileName` sniff to `$oo_prefixes`.

The `FileName` sniff also has an `$exclude` property which is used to indicate which file names should be _excluded_ from the strict file content specific part of the `FileName` sniff checks.

As a number of WPCS sniffs also have an `$exclude` property with a different meaning, this property again can be confusing, so we may as well rename that property now as well.

**Note: This is a backward-compatibility break.**

* This means that the next version will need to be YoastCS 2.0.
* Existing custom rulesets using these properties will need to be adjusted to reflect the change.