:sparkles: New `ValidHookName` sniff
This new sniff extends the upstream WPCS `ValidHookName` sniff and overloads the `transform()` method to allow for the new Yoast specific hook prefix requirements as per issue 143.

Hooks in the Yoast plugins should still comply with the "_hook names should be lowercase with words separated by underscores_" rule, but the `Yoast\WP\PluginName` prefix should be exempt.

The overload of the `transform()` method prevents the upstream sniff from throwing errors for the prefix, while still allowing it to throw errors for the actual hook name.

Also, and in contrast to the WPCS native sniff, when the YoastCS version of this sniff has received valid prefix(es), the hook name checks will only be executed for hooks which have a valid prefix.
Non-prefixed hooks or hooks with another prefix will be ignored to prevent errors being thrown about hook names which are outside of our control.

The sniff uses the `CustomPrefixesTrait` to allow setting the `public` `$prefixes` property and to validate the input received for this property.

In addition to this, the new/extended sniff adds a couple of new checks:
- A check that the "namespace"-style prefix is used.
    The sniff allows for passing both "namespace"-like, as well as "old-style" prefixes to allow for the transition period.
    However, if an "old-style" prefix is used, rather than the "namespace"-like prefix, a `warning` will be thrown.
- A check to verify that hook names contain no more than four words / three underscores, after the plugin specific prefix.
    - The sniff will throw a `warning` when the name consists of more words than `soft_maximum_depth` and an `error` when the name is longer than `maximum_depth`.
    - Both `soft_maximum_depth`, as well as `maximum_depth` are configurable and can be changed from a custom ruleset.
        By default, `soft_maximum_depth` and `maximum_depth` are set to the same value: `4`, i.e. four words.

Note:
For dynamic hook names where the hook name length can not reliably be determined, the sniff will throw a `warning` at severity `3` suggesting the hook name be inspected manually.

As the default severity for PHPCS is `5`, this `warning` at severity `3` will normally not be shown. It is there to allow for intermittently checking of the dynamic hook names. To trigger it, `--severity=3` should be passed on the command line.

Also note:
Just like the lowercase/underscore hook name checks, the hook name length check will only be executed when the hook name starts with (one of) the plugin specific prefix.

Includes unit tests.
Includes documentation.

Fixes 143