:sparkles: New `CustomPrefixes` Trait
This trait can be used by sniffs which need access to the custom plugin prefixes.

The trait contains the `public` property to allow for setting the `$prefixes` from within a custom ruleset, as well as a couple of utility methods to validate and clean the received `$prefixes` so they are usable by the sniffs.

The trait is tested via the sniffs using it.