:sparkles: New `NamespaceName` sniff
This new sniff checks that the names of declared namespaces contain no more than three levels after the plugin specific prefix.

The sniff will throw a `warning` when the name consists of more levels than `recommended_max_levels` and an `error` when the name is longer than `max_levels`.

Both `recommended_max_levels`, as well as `max_levels` are configurable and can be changed from a custom ruleset.

By default, `recommended_max_levels` and `max_levels` are set to respectively 2 and 3.

For the purposes of this sniff, `Tests\` and `Tests\Doubles\` in the namespace name will be disregarded for this level depth check, but only when these are found directly after the prefix.

Additionally, the sniff checks that the levels in the namespace name reflect the directory path of the file based on the project root as identified via the `basepath` setting.

Additional `root` directories can be indicated by setting the `$src_directory` property.

The `CustomPrefixes` trait is used to allow for setting and validating the plugin specific prefix to take into account.

Includes unit tests.
Includes documentation.

Fixes 141