ValidHookName: minor tweak to tests
This minor tweak couldn't be pulled in the ValidHookName sniff PR as at that time, YoastCS still supported WPCS < 2.2.0.

WPCS 2.2.0 contains a bug fix to ignore string array keys which are part of a hook name.

By adjusting this unit test, we safeguard that the YoastCS version of the sniff doesn't undo that fix.