Composer: update minimum PHPCS version
PHP_CodeSniffer has released vesion `3.5.2` in the mean time.

The `3.5.0` release contained a lot of new sniffs, mostly related to PSR12 and a number of which we may want to use for YoastCS.

Sniff additions to the ruleset will be pulled separately.
Updating the minimum supported PHPCS version will allow for that.

Refs:
* https://github.com/squizlabs/PHP_CodeSniffer/releases/tag/3.5.0
* https://github.com/squizlabs/PHP_CodeSniffer/releases/tag/3.5.1
* https://github.com/squizlabs/PHP_CodeSniffer/releases/tag/3.5.2