ObjectNameDepth: allow for extra word for test/double/mock classes
The new `ObjectNameDepth` sniff has a hard limit of "3 words".

Unit test classes often reflect the name of the class being tested + _Test.
Example:
Class `Post_Data_Service` would be tested in a test class called `Post_Data_Service_Test`.

Along the same vein, mock/double classes often reflect the name of the original class with a `_Mock` or `_Double` suffix.

That meant that for plugins with unit tests which adhere to that practice, the word limit effectively was 2 instead of 3 as otherwise they run into problems with the naming of the unit test classes.

This PR adjusts the sniff to take unit tests and test doubles/mocks into account.

For classes with a `_Mock` or `_Double` suffix, the sniff checks that this is an `extend`ed class and if so, will disregard the suffix when determining the word length of the class name.

For classes with a `_Test` suffix, the sniff will check that the class extends a known unit test class and if so, will disregard the suffix when determining the word length of the class name.

Includes unit tests.
Includes updated documentation.

Also:
* Updated the version nrs for the sniff to be in line with the release in which this new sniff will be contained.