Ruleset: disallow Yoda conditions
As proposed in 166 and possible now the minimum PHPCS version is 3.5.0, this adds a sniff to disallow Yoda conditions to YoastCS.

Yoda conditions are enforced by WPCS, but this rule has been turned off in YoastCS from the very beginning, however, non-Yoda conditions were not enforced up to now, leading to inconsistency.

Prior to PHPCS 3.5.2 this sniff was still quite buggy, but it looks like most bugs have been addressed in the mean time.

Fixes 166