Ruleset: check spacing around spread operator
PHPCS includes a new sniff to check that the spacing between the spread operator and the variable/function call it applies to is consistent.

The default behaviour of this sniff is to not allow whitespace between the spread operator and next token.

It is expected that this sniff will eventually be added to WPCS as well.

Refs:
* squizlabs/PHP_CodeSniffer 2548
* WordPress/WordPress-Coding-Standards 1762