Ruleset: disallow "// end" comments
In the olden days, before there were good IDEs, it was often customary to have `// End class`/`// End function` etc comments on the line of a class/function/control structure closing brace to make its easier to find the end of such a structure.

As modern IDEs are perfectly capable of handling this and handle this well, those type of comments are no longer necessary and can be considered unnecessary "noise".

The sniff which is proposed to be added via this PR enforces that closing braces of classes/interfaces/traits/functions are not followed by a comment or statement, in line with PSR-12.