NamingConventions/NamespaceName: implement use of the PHPCSUtils `Namespaces::getDeclaredName()` method
... to replace a large chunk of code.

This also fixes a bug where the original code did not correctly determine the namespace name if it contained a parse error, while the PHPCSUtils method does handle this more robustly.

As a consequence of this change, one test expectation changes, but as that test is for a parse error, I'm not concerned about that.