Composer: use PHPCompatibility 10.0.0-alpha
PHPCompatibility has, at long last, tagged a 10.0 release, even though still in alpha.
PHPCompatibility 10.0.0-alpha1 is cross-version compatible with PHPCS 3.x as well as 4.x.

We really should upgrade to benefit from the > 50 new sniffs and other improvements in this version.

The PHPCompatibilityWP package has received a corresponding 3.0.0-alpha1 tag.

Note: for the plugins to be able to install with these dev versions, they need to have `minimum-stability: dev` (or `alpha`) enabled in their `composer.json`.

Refs:
* https://github.com/PHPCompatibility/PHPCompatibility/wiki/Upgrading-to-PHPCompatibility-10.0
* https://github.com/PHPCompatibility/PHPCompatibilityWP/releases/tag/3.0.0-alpha1
* https://github.com/PHPCompatibility/PHPCompatibility/releases/tag/10.0.0-alpha1